+++ src/cmdTemplateDelete.go
@@ -11,7 +11,16 @@ type templateDeleteCmd struct {
 	AerospikeVersion TypeAerospikeVersion `short:"v" long:"aerospike-version" description:"Aerospike server version (or 'all')"`
 	DistroName       TypeDistro           `short:"d" long:"distro" description:"Linux distro, one of: ubuntu|centos|amazon (or 'all')"`
 	DistroVersion    TypeDistroVersion    `short:"i" long:"distro-version" description:"ubuntu:22.04|20.04|18.04 centos:8|7 amazon:2 (or 'all')"`
+	Aws              templateDeleteCmdAws `no-flag:"true"`
 	Help             helpCmd              `command:"help" subcommands-optional:"true" description:"Print help"`
+}
+
+type templateDeleteCmdAws struct {
+	IsArm bool `long:"arm" description:"indicate installing on an arm instance"`
+}
+
+func init() {
+	addBackendSwitch("template.destroy", "aws", &a.opts.Template.Delete.Aws)
 }
 
 func (c *templateDeleteCmd) Execute(args []string) error {
@@ -20,6 +29,13 @@ func (c *templateDeleteCmd) Execute(args []string) error {
 	}
 
 	log.Print("Running template.destroy")
+	isArm := c.Aws.IsArm
+	if b.Arch() == TypeArchAmd {
+		isArm = false
+	}
+	if b.Arch() == TypeArchArm {
+		isArm = true
+	}
 	// check template exists
 	versions, err := b.ListTemplates()
 	if err != nil {
@@ -27,7 +43,7 @@ func (c *templateDeleteCmd) Execute(args []string) error {
 	}
 
 	if c.DistroName != "all" && c.DistroVersion != "all" && c.AerospikeVersion != "all" {
-		v := backendVersion{c.DistroName.String(), c.DistroVersion.String(), c.AerospikeVersion.String()}
+		v := backendVersion{c.DistroName.String(), c.DistroVersion.String(), c.AerospikeVersion.String(), isArm}
 
 		inSlice, err := inslice.Reflect(versions, v, 1)
 		if err != nil {
@@ -49,6 +65,9 @@ func (c *templateDeleteCmd) Execute(args []string) error {
 
 	var nerr error
 	for _, v := range versions {
+		if v.isArm != isArm {
+			continue
+		}
 		if c.DistroName == "all" || c.DistroName.String() == v.distroName {
 			if c.DistroVersion == "all" || c.DistroVersion.String() == v.distroVersion {
 				if c.AerospikeVersion == "all" || c.AerospikeVersion.String() == v.aerospikeVersion {